+++ ouzel/graphics/Commands.hpp
@@ -68,7 +68,7 @@ namespace ouzel
         class InitRenderTargetCommand: public Command
         {
         public:
-            InitRenderTargetCommand(uint64_t initRenderTarget,
+            InitRenderTargetCommand(uintptr_t initRenderTarget,
                                     uint32_t initWidth,
                                     uint32_t initHeight,
                                     uint32_t initSampleCount,
@@ -82,7 +82,7 @@ namespace ouzel
             {
             }
 
-            uint64_t renderTarget;
+            uintptr_t renderTarget;
             uint32_t width;
             uint32_t height;
             uint32_t sampleCount;
@@ -92,19 +92,19 @@ namespace ouzel
         class DeleteRenderTargetCommand: public Command
         {
         public:
-            explicit DeleteRenderTargetCommand(uint64_t initRenderTarget):
+            explicit DeleteRenderTargetCommand(uintptr_t initRenderTarget):
                 Command(Command::Type::DELETE_RENDER_TARGET),
                 renderTarget(initRenderTarget)
             {
             }
 
-            uint64_t renderTarget;
+            uintptr_t renderTarget;
         };
 
         class SetRenderTargetParametersCommand: public Command
         {
         public:
-            SetRenderTargetParametersCommand(uint64_t initRenderTarget,
+            SetRenderTargetParametersCommand(uintptr_t initRenderTarget,
                                              bool initClearColorBuffer,
                                              bool initClearDepthBuffer,
                                              Color initClearColor,
@@ -118,7 +118,7 @@ namespace ouzel
             {
             }
 
-            uint64_t renderTarget;
+            uintptr_t renderTarget;
             bool clearColorBuffer;
             bool clearDepthBuffer;
             Color clearColor;
@@ -128,38 +128,38 @@ namespace ouzel
         class SetRenderTargetCommand: public Command
         {
         public:
-            explicit SetRenderTargetCommand(uint64_t initRenderTarget):
+            explicit SetRenderTargetCommand(uintptr_t initRenderTarget):
                 Command(Command::Type::SET_RENDER_TARGET),
                 renderTarget(initRenderTarget)
             {
             }
 
-            uint64_t renderTarget;
+            uintptr_t renderTarget;
         };
 
         class ClearRenderTargetCommand: public Command
         {
         public:
-            explicit ClearRenderTargetCommand(uint64_t initRenderTarget):
+            explicit ClearRenderTargetCommand(uintptr_t initRenderTarget):
                 Command(Command::Type::CLEAR_RENDER_TARGET),
                 renderTarget(initRenderTarget)
             {
             }
 
-            uint64_t renderTarget;
+            uintptr_t renderTarget;
         };
 
         // TODO: implement
         class BlitCommand: public Command
         {
         public:
-            BlitCommand(uint64_t initSourceTexture,
+            BlitCommand(uintptr_t initSourceTexture,
                         uint32_t initSourceLevel,
                         uint32_t initSourceX,
                         uint32_t initSourceY,
                         uint32_t initSourceWidth,
                         uint32_t initSourceHeight,
-                        uint64_t initDestinationTexture,
+                        uintptr_t initDestinationTexture,
                         uint32_t initDestinationLevel,
                         uint32_t initDestinationX,
                         uint32_t initDestinationY):
@@ -177,13 +177,13 @@ namespace ouzel
             {
             }
 
-            uint64_t sourceTexture;
+            uintptr_t sourceTexture;
             uint32_t sourceLevel;
             uint32_t sourceX;
             uint32_t sourceY;
             uint32_t sourceWidth;
             uint32_t sourceHeight;
-            uint64_t destinationTexture;
+            uintptr_t destinationTexture;
             uint32_t destinationLevel;
             uint32_t destinationX;
             uint32_t destinationY;
@@ -193,13 +193,13 @@ namespace ouzel
         class ComputeCommand: public Command
         {
         public:
-            explicit ComputeCommand(uint64_t initShader):
+            explicit ComputeCommand(uintptr_t initShader):
                 Command(Command::Type::COMPUTE),
                 shader(initShader)
             {
             }
 
-            uint64_t shader;
+            uintptr_t shader;
         };
 
         class SetCullModeCommad: public Command
@@ -271,25 +271,25 @@ namespace ouzel
         class SetPipelineStateCommand: public Command
         {
         public:
-            SetPipelineStateCommand(uint64_t initBlendState,
-                                    uint64_t initShader):
+            SetPipelineStateCommand(uintptr_t initBlendState,
+                                    uintptr_t initShader):
                 Command(Command::Type::SET_PIPELINE_STATE),
                 blendState(initBlendState),
                 shader(initShader)
             {
             }
 
-            uint64_t blendState;
-            uint64_t shader;
+            uintptr_t blendState;
+            uintptr_t shader;
         };
 
         class DrawCommand: public Command
         {
         public:
-            DrawCommand(uint64_t initIndexBuffer,
+            DrawCommand(uintptr_t initIndexBuffer,
                         uint32_t initIndexCount,
                         uint32_t initIndexSize,
-                        uint64_t initVertexBuffer,
+                        uintptr_t initVertexBuffer,
                         Renderer::DrawMode initDrawMode,
                         uint32_t initStartIndex):
                 Command(Command::Type::DRAW),
@@ -302,10 +302,10 @@ namespace ouzel
             {
             }
 
-            uint64_t indexBuffer;
+            uintptr_t indexBuffer;
             uint32_t indexCount;
             uint32_t indexSize;
-            uint64_t vertexBuffer;
+            uintptr_t vertexBuffer;
             Renderer::DrawMode drawMode;
             uint32_t startIndex;
         };
@@ -334,7 +334,7 @@ namespace ouzel
         class InitBlendStateCommand: public Command
         {
         public:
-            InitBlendStateCommand(uint64_t initBlendState,
+            InitBlendStateCommand(uintptr_t initBlendState,
                                   bool initEnableBlending,
                                   BlendState::Factor initColorBlendSource, BlendState::Factor initColorBlendDest,
                                   BlendState::Operation initColorOperation,
@@ -354,7 +354,7 @@ namespace ouzel
             {
             }
 
-            uint64_t blendState;
+            uintptr_t blendState;
             bool enableBlending;
             BlendState::Factor colorBlendSource;
             BlendState::Factor colorBlendDest;
@@ -368,19 +368,19 @@ namespace ouzel
         class DeleteBlendStateCommand: public Command
         {
         public:
-            DeleteBlendStateCommand(uint64_t initBlendState):
+            DeleteBlendStateCommand(uintptr_t initBlendState):
                 Command(Command::Type::DELETE_BLEND_STATE),
                 blendState(initBlendState)
             {
             }
 
-            uint64_t blendState;
+            uintptr_t blendState;
         };
 
         class InitBufferCommand: public Command
         {
         public:
-            InitBufferCommand(uint64_t initBuffer,
+            InitBufferCommand(uintptr_t initBuffer,
                               Buffer::Usage initUsage,
                               uint32_t initFlags,
                               const std::vector<uint8_t>& initData,
@@ -394,7 +394,7 @@ namespace ouzel
             {
             }
 
-            uint64_t buffer;
+            uintptr_t buffer;
             Buffer::Usage usage;
             uint32_t flags;
             std::vector<uint8_t> data;
@@ -404,19 +404,19 @@ namespace ouzel
         class DeleteBufferCommand: public Command
         {
         public:
-            DeleteBufferCommand(uint64_t initBuffer):
+            DeleteBufferCommand(uintptr_t initBuffer):
                 Command(Command::Type::DELETE_BUFFER),
                 buffer(initBuffer)
             {
             }
 
-            uint64_t buffer;
+            uintptr_t buffer;
         };
 
         class SetBufferDataCommand: public Command
         {
         public:
-            SetBufferDataCommand(uint64_t initBuffer,
+            SetBufferDataCommand(uintptr_t initBuffer,
                                  const std::vector<uint8_t>& initData):
                 Command(Command::Type::SET_BUFFER_DATA),
                 buffer(initBuffer),
@@ -424,14 +424,14 @@ namespace ouzel
             {
             }
 
-            uint64_t buffer;
+            uintptr_t buffer;
             std::vector<uint8_t> data;
         };
 
         class InitShaderCommand: public Command
         {
         public:
-            InitShaderCommand(uint64_t initShader,
+            InitShaderCommand(uintptr_t initShader,
                               const std::vector<uint8_t>& initFragmentShader,
                               const std::vector<uint8_t>& initVertexShader,
                               const std::set<Vertex::Attribute::Usage>& initVertexAttributes,
@@ -455,7 +455,7 @@ namespace ouzel
             {
             }
 
-            uint64_t shader;
+            uintptr_t shader;
             std::vector<uint8_t> fragmentShader;
             std::vector<uint8_t> vertexShader;
             std::set<Vertex::Attribute::Usage> vertexAttributes;
@@ -470,13 +470,13 @@ namespace ouzel
         class DeleteShaderCommand: public Command
         {
         public:
-            DeleteShaderCommand(uint64_t initShader):
+            DeleteShaderCommand(uintptr_t initShader):
                 Command(Command::Type::DELETE_SHADER),
                 shader(initShader)
             {
             }
 
-            uint64_t shader;
+            uintptr_t shader;
         };
 
         class SetShaderConstantsCommand: public Command
@@ -497,7 +497,7 @@ namespace ouzel
         class InitTextureCommand: public Command
         {
         public:
-            InitTextureCommand(uint64_t initTexture,
+            InitTextureCommand(uintptr_t initTexture,
                                const std::vector<Texture::Level>& initLevels,
                                uint32_t initFlags,
                                uint32_t initSampleCount,
@@ -511,7 +511,7 @@ namespace ouzel
             {
             }
 
-            uint64_t texture;
+            uintptr_t texture;
             std::vector<Texture::Level> levels;
             uint32_t flags;
             uint32_t sampleCount;
@@ -521,19 +521,19 @@ namespace ouzel
         class DeleteTextureCommand: public Command
         {
         public:
-            DeleteTextureCommand(uint64_t initTexture):
+            DeleteTextureCommand(uintptr_t initTexture):
                 Command(Command::Type::DELETE_TEXTURE),
                 texture(initTexture)
             {
             }
 
-            uint64_t texture;
+            uintptr_t texture;
         };
 
         class SetTextureDataCommand: public Command
         {
         public:
-            SetTextureDataCommand(uint64_t initTexture,
+            SetTextureDataCommand(uintptr_t initTexture,
                                   const std::vector<Texture::Level>& initLevels):
                 Command(Command::Type::SET_TEXTURE_DATA),
                 texture(initTexture),
@@ -541,14 +541,14 @@ namespace ouzel
             {
             }
 
-            uint64_t texture;
+            uintptr_t texture;
             std::vector<Texture::Level> levels;
         };
 
         class SetTextureParametersCommand: public Command
         {
         public:
-            SetTextureParametersCommand(uint64_t initTexture,
+            SetTextureParametersCommand(uintptr_t initTexture,
                                         Texture::Filter initFilter,
                                         Texture::Address initAddressX,
                                         Texture::Address initAddressY,
@@ -562,7 +562,7 @@ namespace ouzel
             {
             }
 
-            uint64_t texture;
+            uintptr_t texture;
             Texture::Filter filter;
             Texture::Address addressX;
             Texture::Address addressY;
@@ -572,14 +572,14 @@ namespace ouzel
         class SetTexturesCommand: public Command
         {
         public:
-            explicit SetTexturesCommand(uint64_t initTextures[Texture::LAYERS]):
+            explicit SetTexturesCommand(uintptr_t initTextures[Texture::LAYERS]):
                 Command(Command::Type::SET_TEXTURES)
             {
                 for (uint32_t i = 0; i < Texture::LAYERS; ++i)
                     textures[i] = initTextures[i];
             }
 
-            uint64_t textures[Texture::LAYERS];
+            uintptr_t textures[Texture::LAYERS];
         };
     } // namespace graphics
 } // namespace ouzel
